Refactoring
Login and Registration form refactored - more readable.

Removed the unnecessary components (useFormLog and useFormReg)
